requirements.txt
# -----------------------------
# Core runtime (Python 3.11)
# -----------------------------

typing-extensions>=4.8.0,<5.0.0

# -----------------------------
# Orchestration
# -----------------------------

prefect>=2.14.0,<3.0.0

# -----------------------------
# Networking & HTTP
# -----------------------------

requests>=2.31.0,<3.0.0
idna>=3.4,<4.0
charset-normalizer>=3.3.0,<4.0
chardet>=5.2.0,<6.0

# -----------------------------
# DNS / Domain Intelligence
# -----------------------------

dnspython>=2.6.0,<3.0.0
python-whois>=0.9.4,<1.0.0
python-dateutil>=2.9.0,<3.0.0

# -----------------------------
# Testing
# -----------------------------

pytest>=8.0.0,<9.0.0
pytest-mock>=3.12.0,<4.0.0

# -----------------------------
# Optional (LLM explanation layer)
# Uncomment only if CrewAI is used
# -----------------------------

# crewai>=0.28.0,<1.0.0
# langchain>=0.1.0,<0.2.0
